Move mock to unit test file
