--- internal/seller/mocks/mock.go
@@ -1,42 +1,0 @@
-package mocks
-
-import (
-	"context"
-
-	"github.com/extmatperez/meli_bootcamp_go_w2-4/internal/domain"
-	"github.com/stretchr/testify/mock"
-)
-
-type RepositoryMock struct {
-	mock.Mock
-}
-
-func (r *RepositoryMock) GetAll(ctx context.Context) ([]domain.Seller, error) {
-	args := r.Called(ctx)
-	return args.Get(0).([]domain.Seller), args.Error(1)
-}
-
-func (r *RepositoryMock) Get(ctx context.Context, id int) (domain.Seller, error) {
-	args := r.Called(ctx, id)
-	return args.Get(0).(domain.Seller), args.Error(1)
-}
-
-func (r *RepositoryMock) Exists(ctx context.Context, cid int) bool {
-	args := r.Called(ctx, cid)
-	return args.Get(0).(bool)
-}
-
-func (r *RepositoryMock) Save(ctx context.Context, s domain.Seller) (int, error) {
-	args := r.Called(ctx, s)
-	return args.Get(0).(int), args.Error(1)
-}
-
-func (r *RepositoryMock) Update(ctx context.Context, s domain.Seller) error {
-	args := r.Called(ctx, s)
-	return args.Error(1)
-}
-
-func (r *RepositoryMock) Delete(ctx context.Context, id int) error {
-	args := r.Called(ctx, id)
-	return args.Error(1)
-}
